demo/pyproject.toml
[project]
name = "pca-face-recognition-demo"
version = "0.1.0"
description = "Interactive PCA Face Recognition Demo for Master's Academic Evaluation"
authors = [
    {name = "PCA Face Recognition Team", email = "team@example.com"}
]
readme = "docs/us-001-pca-facial-recognition.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12"
]
dependencies = [
    "streamlit>=1.28",
    "numpy>=1.24",
    "scipy>=1.10",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "plotly>=5.15",
    "pillow>=10.0",
    "scikit-learn>=1.3",
    "opencv-python>=4.8",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "streamlit-antd-components (>=0.3.2,<0.4.0)"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "black>=23.9",
    "flake8>=6.1",
    "mypy>=1.6"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include = "src"}]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]